requirements.txt
anthropic>=0.5.0
gitpython>=3.1.30
python-dotenv>=1.0.0
requests>=2.28.2
typer>=0.9.0
rich>=13.3.5
